Small changes to import persons
On line 91 the add function caused $resultArray to add new array objects to it. This would not work with line 112 because it expected an array with objects items with property hr2d__Employee__c.

Line 94 to 98 where redundant. This piece of code is a copy of what is in the do{}until loop. I changed the value of the until from 0 to -1. 

Line 108 to 111 was not working with the correct dataset. It was working with $arbeidsRelatieData instead of the resultArray.